+++ services/horizon/internal/integration/state_verifier_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/test/integration"
 	"github.com/stellar/go/txnbuild"
+	"github.com/stellar/go/xdr"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -63,7 +64,7 @@ func TestStateVerifier(t *testing.T) {
 			Selling:       txnbuild.NativeAsset{},
 			Buying:        txnbuild.CreditAsset{"ABCD", master.Address()},
 			Amount:        "3",
-			Price:         "1",
+			Price:         xdr.Price{1, 1},
 		},
 		&txnbuild.ManageData{
 			SourceAccount: sponsoredSource.AccountID,